Start patching Testem.emitMessage to detect testem events
DOM scraping halts after certain number of calls. This approach patches
emitMessage to send events to console. The other part listens for
console messages and detect when tests start/finish running. There is
still a problem of waiting a little bit after that last test. Not 100%
why is that yet.